refactor(Dao): Dao 인터페이스 구현
